# swagger.yml
swagger: "2.0"
info:
  description: "This is the API documentation for the android application WGPlaner."
  version: "0.0.1"
  title: "WGPlaner API Documentation"
  contact:
    name: WGPlaner-Team
    email: "wgplaner@ameyering.de"
  license:
    name: "MIT"
    url: "https://github.com/WGPlaner/wg_planer_server/blob/master/LICENSE"
host: "api.wgplaner.ameyering.de"
basePath: "/"
...
schemes:
  - https
consumes:
  - application/json
produces:
  - application/json
securityDefinitions:
  UserIDAuth:
    description: For accessing user related parts of the API a valid userID must be passed in 'Authorization' header.
    type: apiKey
    name: Authorization
    in: header
  FirebaseIDAuth:
    description: For accessing user related parts of the API a valid userID must be passed in 'Authorization' header.
    type: apiKey
    name: Authorization
    in: header
paths:
...
  /users/{userId}:
    get:
      tags:
      - user
      description: Returns the user object
      operationId: getUser
      security:
        - FirebaseIDAuth: []
      parameters:
        - name: userId
          in: path
          description: The internal ID of the user
          required: true
          type: string
      responses:
        200:
          description: Success
          schema:
            $ref: "#/definitions/User"
        400:
          description: Invalid userId
          schema:
            $ref: "#/definitions/ErrorResponse"
        401:
          description: Unauthorized User
          schema:
            $ref: "#/definitions/ErrorResponse"
        404:
          description: User not found on server
          schema:
            $ref: "#/definitions/ErrorResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /users/{userId}/image:
    get:
      tags:
      - user
      description: Get the user's profile image
      operationId: getUserImage
      security:
        - UserIDAuth: []
      parameters:
        - name: userId
          in: path
          description: The internal ID of the user
          required: true
          type: string
      produces:
        - application/octet-stream
        - application/json
      responses:
        200:
          description: Success
          schema:
            type: string
            format: binary
            readOnly: true
        400:
          description: Invalid userId
          schema:
            $ref: "#/definitions/ErrorResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    put:
      tags:
      - user
      description: Update the user's profile image
      operationId: updateUserImage
      security:
        - UserIDAuth: []
      parameters:
        - name: userId
          in: path
          description: The internal ID of the user
          required: true
          type: string
        - name: profileImage
          in: formData
          description: Success
          required: true
          type: file
      consumes:
        - multipart/form-data
      responses:
        200:
          description: Success
          schema:
            $ref: "#/definitions/SuccessResponse"
        400:
          description: Invalid userId
          schema:
            $ref: "#/definitions/ErrorResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /users:
    post:
      tags:
      - user
      description: Creates a new user
      operationId: createUser
      security:
        - FirebaseIDAuth: []
      parameters:
      - in: body
        name: body
        description: The name of the user to create
        required: true
        schema:
          $ref: "#/definitions/User"
      responses:
        200:
          description: Success
          schema:
            $ref: "#/definitions/User"
        400:
          description: Invalid user data
          schema:
            $ref: "#/definitions/ErrorResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    put:
      tags:
      - user
      description: Update an existing user
      operationId: updateUser
      security:
        - UserIDAuth: []
      parameters:
      - in: body
        name: body
        description: The users data to update
        required: true
        schema:
          $ref: "#/definitions/User"
      responses:
        200:
          description: Success
          schema:
            $ref: "#/definitions/User"
        400:
          description: Invalid user data
          schema:
            $ref: "#/definitions/ErrorResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
definitions:
  User:
    required:
      - uid
      - displayName
    properties:
      uid:
        type: string
      displayName:
        type: string
        maxLength: 20
        minLength: 3
      groupUid:
        type: string
        format: uuid
      firebaseInstanceId:
        type: string
        pattern: "^[-_:a-zA-Z0-9]{152}$"
      locale:
        type: string
      photoUrl:
        type: string
        format: uri
      email:
        type: string
        format: email
      createdAt:
        type: string
        format: date-time
        readOnly: true
      updatedAt:
        type: string
        format: date-time
        readOnly: true
...
  ErrorResponse:
    required:
      - status
      - message
    properties:
      status:
        type: integer
      message:
        type: string
  SuccessResponse:
    required:
      - status
      - message
    properties:
      status:
        type: integer
      message:
        type: string